requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
-e file:///Users/idmit/Documents/dev/newbox/qzbot
certifi==2017.4.17        # via python-telegram-bot
future==0.16.0            # via python-telegram-bot
python-telegram-bot==6.0.3
